refactor(evaluation): move the PII scorecard beside the Sarvam one
Both scorecards landed on 2026-08-07 from separate Sprint 3 branches that
never saw each other. pii_scorecard went into pipeline/ because it imports
pipeline.pii_eval and sits beside it; sarvam_scorecard has no janasunani
imports at all, so nothing pulled it anywhere and its branch created a new
evaluation/ package. Each was locally reasonable and the two disagree.

The new package's own docstring already contradicted the split: "score
pipeline stages against ground truth" describes pii_scorecard exactly.

Resolved toward evaluation/, and the rule is now written down. The line is
what the code does with a bad number, not what it computes: pii_eval is a
gate the pipeline runs and can fail a release, so it stays in pipeline/
with the thing it gates; these report, nothing imports them, and each has
its own CLI.

The ARCHITECTURE package table was wrong about six of fourteen packages.
It listed egress/ and experiments/ as planned when egress/ has shipped,
and omitted analytics, evaluation, inference, pii, routing and serving
entirely. Every package on disk is now listed; experiments/ stays marked
planned because it genuinely does not exist yet.

ruff clean, 1024 passed, 7 skipped.
